tests/baseline/protocol-1.2/inspector/console/console-clear-function-expected.txt

CONSOLE MESSAGE: line 10: one
CONSOLE MESSAGE: line 11: two
CONSOLE MESSAGE: line 12: three
CONSOLE MESSAGE: line 19: console.clear
Tests that console is cleared via console.clear() method

Bug 101021
=== Before clear ===
console-clear-function.html:10 one
console-clear-function.html:11 two
console-clear-function.html:12 three
=== After clear ===
console-clear-function.html:19 Console was cleared
